extract TransparentIndexer sub-trait into globally_public module
  Move all transparent address methods into a TransparentIndexer: Indexer
  sub-trait with a single TransparentError associated type. The trait
  definition and GrpcIndexer impl live in src/globally_public.rs, gated
  by one #[cfg(feature = "globally-public-transparent")] on the module.

  Also:
  - Add ping-very-insecure feature gate for Ping RPC (name mirrors the
    lightwalletd --ping-very-insecure CLI flag required server-side)
  - Bound all associated error types with std::error::Error
  - Rename GetTreesError to GetTreeStateError to match the proto/method
  - Strengthen doc comments across all trait methods
